Minor reorg
